pagerank and main

+++ pagerank.h
@@ -13,9 +13,10 @@ class PageRank {
         vector<unsigned> get_airport_rank();
         void store_pagerank();
     private:
-        vector<unsigned> rank_;
+        vector<double> rank_;
         vector<Airport> airports_;
         unordered_map<unsigned, vector<unsigned >> adjList;
         map<unsigned, unsigned> mp;
         size_t num_;
+
 };
